Separate split runtime planning concerns
